requirements.txt
# Database
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.4
numpy==1.26.2

# API requests
requests==2.31.0

# Configuration
PyYAML==6.0.1

# Logging (built-in, but explicit)
# logging - built-in
